Bug 2043518: set degraded and available status based on Prometheus pod status
This patch sets the Degraded and Availability status of ClusterOperator
by inspecting Prometheus Pod status. To achieve this all tasks now
return a StateError that captures if the state of the system is
Degraded or Unavailable and its  associated reasons for the change.

Thus a task's Run method is now structured as follows

```go

func (t *Task) Run() *StateError {
  creationError := t.create()
  validateError := t.validate()
  return mergeStateErrors(creationError, validationErr)
}
```

This patch only modifies the Prometheus Task to add a validation step
but can be applied to other tasks as well.

Thus, given the following monitoring config which uses a non-existent PVC

```yaml
  prometheusK8s:
    volumeClaimTemplate:
      metadata:
        name: prometheus-db
      spec:
        storageClassName: foo
        resources: { requests: {storage: 100Mi} }
```

It now correctly marks the ClusterOperator status as follows (formatted
for readability)

```
  Monitoring Stack is Unavailable.
  Error: updating prometheus-k8s
  status: unavailable, degraded;
  reasons:
    prometheus pod openshift-monitoring/prometheus-k8s-0 is pending due to Unschedulable
  : 0/6 nodes are available
  : 6 pod has unbound immediate PersistentVolumeClaims., prometheus pod openshift-monitoring/prometheus-k8s-1 is pending due to Unschedulable
  : 0/6 nodes are available
  : 6 pod has unbound immediate PersistentVolumeClaims
```

Signed-off-by: Sunil Thaha <sthaha@redhat.com>